moved both sub endpoints to the same file
